Organise the classes into their relevant packages
