+++ src/main/java/Edith/command/Command.java
@@ -1,3 +1,5 @@
+package command;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws EdithException;
