+++ src/main/java/Edith/command/DeleteCommand.java
@@ -1,3 +1,5 @@
+package command;
+
 import java.io.IOException;
 
 public class DeleteCommand extends Command {
@@ -10,7 +12,7 @@ public class DeleteCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws EdithException {
         if (index < 0 || index >= tasks.getNumOfTasks()) {
-            throw new EdithException("Task " + index + " does not exist. Please enter a valid task number.");
+            throw new EdithException("task.Task " + index + " does not exist. Please enter a valid task number.");
         }
 
         Task taskToDelete = tasks.getTask(index);
