+++ src/main/java/Edith/command/ExitCommand.java
@@ -1,3 +1,5 @@
+package command;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws EdithException {
