+++ src/main/java/Edith/command/ListCommand.java
@@ -1,3 +1,5 @@
+package command;
+
 public class ListCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws EdithException {
