+++ src/main/java/Edith/command/MarkCommand.java
@@ -1,23 +1,25 @@
+package command;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     private int index;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws EdithException {
         if (index < 0 || index >= tasks.getNumOfTasks()) {
-            throw new EdithException("Task " + index + " does not exist. Please enter a valid task number.");
+            throw new EdithException("task.Task " + index + " does not exist. Please enter a valid task number.");
         }
 
-        Task taskToUnmark = tasks.getTask(index);
-        taskToUnmark.markTaskUndone();
+        Task taskToMark = tasks.getTask(index);
+        taskToMark.markTaskDone();
 
-        ui.showIndentedMessage("Sure, I've marked task " + (index + 1) + " as not done yet:");
-        ui.showIndentedMessage(taskToUnmark.toString());
+        ui.showIndentedMessage("Alright, great job! I've marked task " + (index + 1) + " as done:");
+        ui.showIndentedMessage(taskToMark.toString());
         ui.showLineBreak();
 
         try {
