+++ src/main/java/Edith/task/Deadline.java
@@ -1,3 +1,5 @@
+package task;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
