+++ src/main/java/Edith/task/Task.java
@@ -1,3 +1,5 @@
+package task;
+
 public abstract class Task {
     private String typeOfTask;
     private String task;
